Separate classes and interfaces in files
