+++ src/Nethermind/Nethermind.Tools.Kute/IAuth.cs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+namespace Nethermind.Tools.Kute;
+
+interface IAuth
+{
+    string AuthToken { get; }
+}
