+++ src/Nethermind/Nethermind.Tools.Kute/NullJsonRpcSubmitter.cs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+namespace Nethermind.Tools.Kute;
+
+class NullJsonRpcSubmitter : IJsonRpcSubmitter
+{
+    public Task Submit(string content) => Task.CompletedTask;
+}
